refactor(global-state): add api to module
